+++ spec/integration/rack_3_0/version_spec.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+describe Rack do
+  it { expect(Gem::Version.new(described_class.release).segments.first).to eq 3 }
+end